requirements.txt
# Sentinet Dependencies
# =====================

# SDN Controller Framework
ryu>=4.34

# Network Utilities
eventlet>=0.30.0

# AI/ML (optional - for trained models)
scikit-learn>=1.0.0
joblib>=1.1.0
numpy>=1.21.0

# WebSocket Client (for backend communication)
websocket-client>=1.3.0

# Utilities
PyYAML>=6.0
